Refactoring time: tables are now archetypes fields.
This is not complete, but no more incomplete than we were before refatoring.
Only exeption: we do not yet offer a select box to choose a backend source, but
because there is only one source so far the old implementation hid the option
to pick one automatically.

Note that there is a bug in Archetypes (all versions up to 1.6.4) that prevents
the ajax data call to work for any content in the portal factory, so the
datatable is empty in that case. But the plan is to make the edit view not show
the table in the first place, only configuration for it.

git-svn-id: http://svn.plone.org/svn/collective/collective.table/trunk@231822 db7f04ef-aaf3-0310-a811-c281ed44c4ad